refactor(transform): split stream-transform.ts + fix ts-taste issues
- Split stream-transform.ts (469 lines) into 3 files: base, oa2ant, ant2oa
- Extract shared ID generators (generateMsgId, generateChatcmplId) to id-utils.ts
- Replace silent catches with console.warn/error logging across 6 files
- Use UPSTREAM_SUCCESS constant instead of magic number 200 in transport-fn
- Remove duplicate eslint-disable directive in plugin-registry
