+++ src/proxy/transport/transport-fn.ts
@@ -102,7 +102,7 @@ export function buildTransportFn(p: TransportFnParams): (target: Target) => Prom
       if (mr) p.tracker?.update(p.logId, { streamMetrics: toStreamMetrics(mr) });
     }
     // Apply format transformation first (e.g. Anthropic→OpenAI), then inject model-info tag
-    if (p.responseTransform && result.kind === "success" && result.statusCode === 200 && result.body) {
+    if (p.responseTransform && result.kind === "success" && result.statusCode === UPSTREAM_SUCCESS && result.body) {
       result = { ...result, body: p.responseTransform(result.body) };
     }
     if (p.responseTransform && (result.kind === "error" || result.kind === "stream_error") && result.body) {
@@ -115,7 +115,7 @@ export function buildTransportFn(p: TransportFnParams): (target: Target) => Prom
           bodyObj.content[0].text += `\n\n${buildModelInfoTag(p.effectiveModel)}`;
           return { ...result, body: JSON.stringify(bodyObj) };
         }
-      } catch { p.request.log.debug("Failed to inject model-info tag into non-JSON response"); }
+      } catch { p.request.log.warn("Failed to inject model-info tag into non-JSON response"); }
     }
     return result;
   };
